Stage 2-5: drivers, dual-core, PIO v2, packaging, provenance
- SDK: GPIO interrupts, PWM, ADC (temp sensor), UART (zero-wire loopback),
  dual-core launch + hardware spinlocks + inter-core FIFO, PIO side-set /
  forward labels / comptime assembly — 26 on-target tests, all HW-verified
- boot2 built from vendored pico-sdk 2.1.0 source (BSD-3-Clause) with
  CRC32-MPEG2 appended at build time; regression-tested byte-identical
  to the golden blob; SVD refresh task documents provenance
- fair size comparison: same-rig blink in all four languages
  (gcc 712 B / Mojo 780 B = clang 780 B / Rust 784 B), pixi run sizes
- `pico` package installable via pixi add --git (pixi-build-mojo backend)
- project renamed to inmojomni; PIO assembler host unit tests; SWD flash
  hardening (--verify + retry at 1 MHz)

+++ src/pico/rtt.mojo
@@ -107,7 +107,7 @@ def write_u32(v: UInt32):
 def write_hex(v: UInt32):
     """Append `0x` + 8 hex digits."""
     var wr = read32(_UP_WROFF)
-    if 10 > _room(wr):
+    if UInt32(10) > _room(wr):
         return
     wr = _put(wr, 48)  # '0'
     wr = _put(wr, 120)  # 'x'
